requirements.txt
RapidSMS==0.19.0
jsonfield>=1.0.0
django-bootstrap3>=5.1.1
django-rosetta>=0.7.4
django-crispy-forms>=1.4.0
django-nose>=1.3
pytz>=2014.10
